+++ cache/set.go
@@ -1,21 +1,30 @@
-package set
+package cache
 
 import "sync"
 
+// Set is a collection of unique items. It should be thread-safe.
 type Set[T comparable] interface {
+	// Add adds an item to the set.
 	Add(item T)
+	// Remove removes an item from the set.
 	Remove(item T)
+	// Has returns true if the item is in the set, false otherwise.
 	Has(item T) bool
+	// Len returns the number of items in the set.
 	Len() int
+	// Clear removes all items from the set.
 	Clear()
+	// ForEach iterates over the items in the set.
 	ForEach(f func(item T))
 }
+
 type set[T comparable] struct {
 	mu    sync.RWMutex
 	items map[T]struct{}
 }
 
-func New[T comparable]() Set[T] {
+// NewSet returns a thread-safe in memory implementation of a set.
+func NewSet[T comparable]() Set[T] {
 	return &set[T]{
 		items: make(map[T]struct{}),
 	}
